Reverted back Manage Employee...
I eliminated the need for the availability table by adding that
information to the employee table.  Altered the manage employee
index.php to effectively gather data.  Also, added the DB build file to
the root of the directory so we don't have to go to D2L to get it.