request class so far
sure it will change a bit. Going to try having the constructor take in
each value from the query and create the object there. If it works
passing those values we should be able to implement that for each other
class and cut some code out of the other files.