js fix
I had to move the js from the employee over to the js file I submitted
earlier. Needed to make another function to put the end times from the
db back into the back when an existing employee is selected.

I will need to do the same for the requests, but I'll do it another
night.