shifts link
Hey Rob, won't really mess with this page till we get a chance to talk.
I don't mind helping, but don't want to mess anything up if you are
further or anything. Just changed the link to the edit page to the shift
page if we are just going to have the create / edit on a single page.