hoover and time drop down sizes
was planning on rewriting on of the js methods to resemble what you did.
right now if you select an employee from the table and they have a day
off you won't get the hover text. Figure I'll wait till I rewrite the
method rather than writing another giant if block for each time field to
check if length is 0