hidden fields it is
yeah, can't do a set call for the value, so going to have to be  hidden
fields unless you know another way to do it. Figure we can just pass the
the $x from the for loop, or go off the session employee / schedule
array seeing she schedules everyone once a week.