updates for schedule
don't think everything is fixed, but wanted to get most of it up before
you worked much more on it.

I changed the shift queries because it is easier getting the shift times
for adding to the total if the value is as I changed it. If this causes
a problem with anything you are doing(or something I didn't notice) I
can change it back. Is possible to just split the 6:00-10:00 format
apart, but seems easy to just split on the _

Shift hourly totals should work except for shifts. Plan on getting that
done next.

Anything under the max weekly hours should be in black text, and
anything over should be in red. It should switch back to black when you
go back under the max hours.

I fixed the min shift requirement a bit. It wasn't working if you set a
start / end time and then put the end time an hour after the start. If
you select a start / end time that is valid and then change end time to
be invalid it should switch it back to the min shift time

if you click create schedule it goes to the validateSchedule page where
the employee shift times are changed to read how she wanted. It could
obviously use a little formatting work, but I think the "basic" part of
it is about done.

since I added the hours it kind of squeezes the page a bit, so let me
know if you think it is a good idea to just ditch the left hand section
of the page.

if you find anything I messed up lmk